% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{path_data}
\alias{path_data}
\title{Build paths from \code{data} directory}
\usage{
path_data(...)
}
\arguments{
\item{...}{Character vector of subdirectories relative to the \code{data}
directory.}
}
\value{
The full path to the \code{data} directory, or (if provided as a
parameter), the path to the subdirectory.
}
\description{
Build paths from \code{data} directory
}
\seealso{
\link{path_trendecon}
}
