% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{path_data_raw}
\alias{path_data_raw}
\title{Build paths from \code{data-raw} directory}
\usage{
path_data_raw(...)
}
\arguments{
\item{...}{Character vector of subdirectories relative to the \code{data-raw}
directory.}
}
\value{
The full path to the \code{data-raw} directory, or (if provided as a
parameter), the path to the subdirectory.
}
\description{
Build paths from \code{data-raw} directory
}
\seealso{
\link{path_trendecon}
}
