% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{path_trendecon}
\alias{path_trendecon}
\title{Build paths from base directory}
\usage{
path_trendecon(...)
}
\arguments{
\item{...}{Character vector of subdirectories relative to the base
directory.}
}
\value{
The full path to the base directory, or (if provided as a
parameter), the path to the subdirectory.
}
\description{
The base directory is obtained by \code{getOption("path_trendecon")}. If the
option is not present, the base directory defaults to the parent directory
of the current working directory. To set the option, run
}
\details{
\preformatted{options(path_trendecon = "~/path/to/base/dir")
}
}
\examples{
path_trendecon("data-raw")

}
\seealso{
\link[R.utils:setOption]{R.utils::setOption}
}
