% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_aggregate_raw.R
\name{proc_aggregate_raw}
\alias{proc_aggregate_raw}
\title{may become obsolete. creates aggregated series from raw data.}
\usage{
proc_aggregate_raw(keyword = "Insolvenz")
}
\description{
may become obsolete. creates aggregated series from raw data.
}
\seealso{
\code{\link[trendecon:aggregate_windows]{trendecon::aggregate_windows()}}
}
