% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_combine_freq.R
\name{proc_combine_freq}
\alias{proc_combine_freq}
\title{\enumerate{
\item bend daily series (which fullfills weekly constraint) so that monthly
values are identical to monthly series
}}
\usage{
proc_combine_freq(keyword = "Insolvenz")
}
\description{
\enumerate{
\item bend daily series (which fullfills weekly constraint) so that monthly
values are identical to monthly series
}
}
