% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_keyword_init.R
\name{proc_keyword_init}
\alias{proc_keyword_init}
\title{Download daily, weekly and monthly Google Trends data for a keyword.}
\usage{
proc_keyword_init(keyword = "Insolvenz", from = "2006-01-01")
}
\arguments{
\item{keyword}{A single keyword to query Google Trends.}

\item{from}{Start of timeframe in YYYY-mm-dd form.}
}
\description{
Downloads daily, weekly and monthly Google Trends data for a keyword
and writes the data to csv files.
}
\details{
By default, the data is stored in
folders \code{data-raw/indicator_raw} and \code{data-raw/indicator}. File suffixes
are \verb{_d} for daily, \verb{_w} for weekly, and \verb{_m} for monthly data.
\itemize{
\item \code{data-raw/indicator_raw} contains time series in time-windows, as returned
by \link[trendecon:ts_gtrends_windows]{trendecon::ts_gtrends_windows}.
\item \code{data-raw/indicator} contains time series aggregated within time-windows
by \link[trendecon:aggregate_windows]{trendecon::aggregate_windows}.
}
}
\examples{
\dontrun{
proc_keyword_init(keyword = "Insolvenz", from = "2006-01-01")
}

}
\seealso{
\link{ts_gtrends_windows}
}
