% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_keyword_latest.R
\name{proc_keyword_latest}
\alias{proc_keyword_latest}
\title{Download latest data}
\usage{
proc_keyword_latest(keyword = "Insolvenz", n_windows = 12)
}
\arguments{
\item{keyword}{A single keyword for which to process the data.}

\item{n_windows}{Number (integer) of time-windows.}
}
\description{
Downloads latest daily, weekly, and monthly data.
}
\section{Daily data}{

Downloads data for the last 90 days in two windows, with the same end
date (today) but start date shifted by one day. File saved as
\verb{\{keyword\}_d_\{today\}.csv} in folder \code{data-raw/indicator_raw}.
}

\section{Weekly data}{

Downloads weekly data for two windows, first window starts 1 year ago,
second window offset by one week, both windows end today. File saved
as \verb{\{keyword\}_w_\{today\}.csv} in folder \code{data-raw/indicator_raw}.
}

\section{Monthly data}{

Downloads monthly data for two windows, first window starts at
2006-01-01, second window offset by  1 month, both windows end today.
File saved as \verb{\{keyword\}_m_\{today\}.csv} in folder
\code{data-raw/indicator_raw}.
}

\seealso{
\link{ts_gtrends_windows}
}
