% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{read_keywords}
\alias{read_keywords}
\title{Read keyword indicator data from disk}
\usage{
read_keywords(keywords, geo = "CH", suffix = "sa", id = NULL)
}
\arguments{
\item{keywords}{A vector of keywords.}

\item{geo}{geographical area.}

\item{suffix}{Suffix in file names, defaults to \code{"sa"}. Common for all
keywords.}

\item{id}{Category id, defaults to \code{NULL}.}
}
\value{
A tibble with columns \code{keyword}, \code{time}, \code{value}.
}
\description{
Reads keyword indicators from csv files from
\verb{/\{base_dir\}/data-raw/indicator/\{keyword\}_\{suffix\}.csv} where \code{{keyword}}
is one of the keywords in parameter \code{keywords}.
}
