% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends.R
\name{ts_gtrends}
\alias{ts_gtrends}
\title{Download Google Trend Data}
\usage{
ts_gtrends(
  keyword = NA,
  category = "0",
  geo = "CH",
  time = "today+5-y",
  retry = 5,
  wait = 5,
  quiet = FALSE
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Contrary to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}

\item{category}{A character vector, listing the categories, defaults to “0”.}
}
\value{
A tibble of time series for the different keywords or categories. If only a single keyword and a single
category are specified, the tibble has columns \emph{time} and \emph{value}. If either more
than one keywords, or more than one category are given, an additional column \emph{id} indicates
either the keyword, or the category.
}
\description{
Wrapper around \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}} that modifies the original function by
a) vectorizing it, b) converting the result to tsboxable tibble and c)
retries if no result is returned.
}
\section{Categories}{

A list with all categories can be obtained using \code{data("categories")}. The function takes the id's as strings,
not the names. For example, "Arts & Entertainment" has to be specified as "2".
}

\examples{

ts_gtrends("Rezession")
}
\seealso{
\code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, \code{browseVignettes("intro")}

\href{https://www.datacareer.ch/blog/analyzing-google-trends-with-r-retrieve-and-plot-with-gtrendsr/}{Online tutorial for analyzing google trends in R}
}
