% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends_index.R
\name{ts_gtrends_index}
\alias{ts_gtrends_index}
\title{Construct Index based on Google Trend keywords.}
\usage{
ts_gtrends_index(
  keyword,
  time = paste("2007-01-01", Sys.Date()),
  ...,
  sadj = TRUE
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Defaults to NA which searches for the whole category. Contrary
to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}

\item{time}{A string specifying the time span of the query. Possible values
  are:

  \describe{ \item{"now 1-H"}{Last hour} \item{"now 4-H"}{Last four hours}
  \item{"now 1-d"}{Last day} \item{"now 7-d"}{Last seven days} \item{"today
  1-m"}{Past 30 days} \item{"today 3-m"}{Past 90 days} \item{"today
  12-m"}{Past 12 months} \item{"today+5-y"}{Last five years (default)}
  \item{"all"}{Since the beginning of Google Trends (2004)} \item{"Y-m-d
  Y-m-d"}{Time span between two dates (ex.: "2010-01-01 2010-04-03")} }}

\item{...}{further arguments, passed to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}.}

\item{sadj}{should the series be seasonally adjusted? Only works if resulting data is monthly.}
}
\description{
Uses the first principal component to construct an index.
}
\examples{

tsbox::ts_plot(ts_gtrends_index(c( "Mango", "Zara")))
}
\seealso{
\code{browseVignettes("intro")}
}
