% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends_mwd.R
\name{ts_gtrends_mwd}
\alias{ts_gtrends_mwd}
\title{Robust and Consistent Time Series from Google Trends}
\usage{
ts_gtrends_mwd(keyword = NA, category = "0", geo = "CH", from = "2006-01-01")
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Defaults to NA which searches for the whole category. Contrary
to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}

\item{category}{A character vector, listing the categories, defaults to
\code{"0"}. Passed to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}}

\item{geo}{A character vector denoting geographic regions for the query,
default is \code{"CH"}. Passed to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}}

\item{from}{A character string specifying the starting date for the obtained time series, defaults to 2006-01-01}
}
\description{
Construct a robust and consistent daily Time Series from Google Trends data.
Daily, weekly and monthly Data is downloaded and consistently aggregated, using the Chow-Lin methodology.
}
\details{
See the working paper for details on the methodology.
}
