% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends_windows.R
\name{ts_gtrends_windows}
\alias{ts_gtrends_windows}
\title{Get approximate daily series over a large window}
\usage{
ts_gtrends_windows(
  keyword = NA,
  category = "0",
  geo = "CH",
  from = "2019-01-01",
  prevent_window_shrinkage = FALSE,
  stepsize = 7,
  windowsize = 80,
  n_windows = 12,
  quiet = FALSE,
  wait = 60,
  retry = 5
)
}
\arguments{
\item{keyword}{keyword(s) to compare}

\item{category}{category/ies to compare}

\item{geo}{region to get trends for}

\item{from}{start of timeframe in YYYY-mm-dd form}

\item{stepsize}{how far to advance the window in days}

\item{quiet}{should progress output be su}

\item{wait}{waiting time until retry}

\item{to}{end of timeframe in YYYY-mm-dd form}
}
\description{
This function downloads
}
