% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{write_keyword}
\alias{write_keyword}
\title{Write csv file for keyword indicator}
\usage{
write_keyword(x, keyword, geo, suffix = "sa")
}
\arguments{
\item{x}{Tibble of data to write to file.}

\item{keyword}{Keyword (character vector) for which to construct the path
to the indicator.}

\item{geo}{geographical area}

\item{suffix}{Character vector for file suffix, defaults to \code{"sa"}.}
}
\description{
Writes csv file for keyword indicator to
\verb{/\{base_dir\}/data-raw/indicator/\{keyword\}_\{suffix\}.csv}.
}
\seealso{
\link{path_keyword}
}
