% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{path_keyword}
\alias{path_keyword}
\title{Build path to indicator data file}
\usage{
path_keyword(keyword, suffix)
}
\arguments{
\item{keyword}{Keyword (character vector) for which to construct the path
to the indicator.}

\item{suffix}{Character vector for file suffix.}
}
\description{
Builds path to indicator files of the form
\verb{/\{base_dir\}/data-raw/indicator/\{keyword\}_\{suffix\}.csv}.
}
\seealso{
\link{path_data_raw}
}
