% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_index.R
\name{proc_index}
\alias{proc_index}
\title{Prepare an index.}
\usage{
proc_index(keywords, geo, index_name)
}
\arguments{
\item{keywords}{A vector of keyword names comprising the index.}

\item{geo}{A character vector denoting geographic region.}

\item{index_name}{The name given to the index.}
}
\description{
Creates or updates an index using data from already downloaded keywords.
}
\details{
This function assumes the data for all keywords exists.
Then it proceeds to go through several steps:
\enumerate{
\item updates all the keywords by downloading newest Google trends data.
\item combines monthly, weekly, and daily frequencies.
\item runs seasonal adjustment steps.
\item combines all the keywords into one index using principal components.
\item writes the produced index to a file
}
}
\examples{
\dontrun{
keywords <- c("Mango", "Zara", "H&M", "PKZ", "Blue Tomato", "Dosenbach", "Schuhe kaufen", "Ochsner Schuhe")
proc_index(keywords, "ch", "clothing")
}
}
