% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_keyword.R
\name{proc_keyword}
\alias{proc_keyword}
\title{Processes the data for a single keyword.}
\usage{
proc_keyword(keyword = "Insolvenz", n_windows = 2)
}
\arguments{
\item{keyword}{A single keyword for which to process the data.}

\item{n_windows}{Number of windows, passed to \link{ts_gtrends_windows}, used
when downloading the latest data.}
}
\description{
Processing includes
\itemize{
\item donwloading the latest data
\item aggregating
\item combining
\item seasonal adjustment
}
}
\seealso{
\link{ts_gtrends_windows}
}
