% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_keyword_init.R
\name{proc_keyword_init}
\alias{proc_keyword_init}
\title{Download daily, weekly and monthly Google Trends data for a keyword.}
\usage{
proc_keyword_init(
  keyword = "Insolvenz",
  geo = "CH",
  from = "2006-01-01",
  indicator_raw = FALSE
)
}
\arguments{
\item{keyword}{A single keyword to query Google Trends.}

\item{geo}{A character vector denoting the geographic region.
Default is "CH".}

\item{from}{Start of timeframe in YYYY-mm-dd form. Should not be changed from
the the default.}

\item{indicator_raw}{store individual downloads. If \code{FALSE}, only the
averages are stored.}
}
\description{
Downloads daily, weekly and monthly Google Trends data for a keyword
and writes the data to csv files.
}
\details{
By default, the data is stored in folders \code{data} and \code{raw}. Each folder
contains a subdirectory of each country.
}
\examples{
\dontrun{
# run once
proc_keyword_init(keyword = "Insolvenz", "AT")
# run every day
proc_keyword_latest(keyword = "Insolvenz", "AT")
}

}
\seealso{
\link{ts_gtrends_windows}
}
