% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_trendecon_ch.R
\name{proc_trendecon_ch}
\alias{proc_trendecon_ch}
\title{Calculate Trendecon Main Indices for Switzeland}
\usage{
proc_trendecon_ch(path = ".", copy_to_data = FALSE)
}
\arguments{
\item{path}{base directory}
}
\description{
Calculate Trendecon Main Indices for Switzeland
}
