% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends.R
\name{ts_gtrends}
\alias{ts_gtrends}
\title{Download Google Trend Data}
\usage{
ts_gtrends(
  keyword,
  category = "0",
  geo = "CH",
  time = "today+5-y",
  retry = 5,
  wait = 5,
  quiet = FALSE
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Contrary to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}

\item{category}{A character vector, listing the categories, defaults to
\code{"0"}. Passed to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}}

\item{geo}{A character vector denoting geographic regions for the query,
default is \code{"CH"}. Passed to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}}

\item{time}{A string specifying the time span of the query. Possible values
  are:

  \describe{ \item{"now 1-H"}{Last hour} \item{"now 4-H"}{Last four hours}
  \item{"now 1-d"}{Last day} \item{"now 7-d"}{Last seven days} \item{"today
  1-m"}{Past 30 days} \item{"today 3-m"}{Past 90 days} \item{"today
  12-m"}{Past 12 months} \item{"today+5-y"}{Last five years (default)}
  \item{"all"}{Since the beginning of Google Trends (2004)} \item{"Y-m-d
  Y-m-d"}{Time span between two dates (ex.: "2010-01-01 2010-04-03")} }}

\item{retry}{Number of attempts, in case the query request does not succeed.}

\item{wait}{Seconds to wait between attempts, where waiting time is \code{attempt * wait}.}

\item{quiet}{If TRUE won't display messages related to server interactions. Default is FALSE.}
}
\value{
A tibble of time series for the different keywords or categories. If only a single keyword and a single
category are specified, the tibble has columns \emph{time} and \emph{value}. If either more
than one keyword, or more than one category are given, an additional column \emph{id} indicates
either the keyword, or the category.
}
\description{
Wrapper around \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}} that modifies the original function by
a) vectorizing it, b) converting the result to tsboxable tibble and c)
retries if no result is returned.
}
\section{Notes}{

\itemize{
\item The function takes the id's as strings, not the names. For example, "Arts & Entertainment" has to be specified as "2".
\item Multiple keywords or multiple categories can be specified, but not both.
}
}

\examples{

x <- ts_gtrends("Rezession")
x
tsbox::ts_plot(x)
}
\seealso{
\code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, \code{browseVignettes("intro")}

\href{https://www.datacareer.ch/blog/analyzing-google-trends-with-r-retrieve-and-plot-with-gtrendsr/}{Online tutorial for analyzing google trends in R}
}
