% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends.R
\name{ts_gtrends}
\alias{ts_gtrends}
\title{Download Google Trend Data}
\usage{
ts_gtrends(
  keyword = NA,
  category = "0",
  geo = "CH",
  time = "today+5-y",
  retry = 5,
  wait = 5,
  quiet = FALSE
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Contrary to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}
}
\description{
Wrapper around \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}} that modifies the original function by
a) vectorizing it, b) converting the result to tsboxable tibble and c)
retries if no result is returned.
}
\examples{

ts_gtrends("Rezession")
}
\seealso{
\code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, \code{browseVignettes("intro")}

\href{https://www.datacareer.ch/blog/analyzing-google-trends-with-r-retrieve-and-plot-with-gtrendsr/}{Online tutorial for analyzing google trends in R}
}
