% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends_index.R
\name{ts_gtrends_index}
\alias{ts_gtrends_index}
\title{Construct Index based on Google Trend keywords.}
\usage{
ts_gtrends_index(
  keyword,
  time = paste("2007-01-01", Sys.Date()),
  ...,
  sadj = TRUE
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Contrary to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}

\item{time}{A string specifying the time span of the query.
Passed to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}.}

\item{...}{further arguments, passed to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}.}

\item{sadj}{should the series be seasonally adjusted? Only works if resulting data is monthly.}
}
\description{
Uses the first principal component to construct an index.
}
\examples{

tsbox::ts_plot(ts_gtrends_index(c( "Mango", "Zara")))
}
\seealso{
\code{browseVignettes("intro")}
}
