% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends_windows.R
\name{ts_gtrends_windows}
\alias{ts_gtrends_windows}
\title{Get approximate daily series over a large window}
\usage{
ts_gtrends_windows(
  keyword = NA,
  category = "0",
  geo = "CH",
  from = "2019-01-01",
  prevent_window_shrinkage = FALSE,
  stepsize = 7,
  windowsize = 80,
  n_windows = 12,
  quiet = FALSE,
  wait = 60,
  retry = 5
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Contrary to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}

\item{category}{A character vector, listing the categories, defaults to “0”.}

\item{geo}{A character vector denoting geographic regions for the query, default is “CH”.}

\item{from}{start of timeframe in YYYY-mm-dd form}

\item{stepsize}{how far to advance the window in days}

\item{quiet}{If TRUE won't display messages related to server interactions. Default is FALSE.}

\item{wait}{Seconds to wait between attempts, where waiting time is \code{attempt * wait}.}

\item{retry}{Number of attempts, in case the query request does not succeed.}

\item{to}{WHAT HAPPENED WITH THIS?}
}
\description{
This function downloads
}
