% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gtrends_windows.R
\name{ts_gtrends_windows}
\alias{ts_gtrends_windows}
\title{Download approximate daily series over a sequence of time-windows}
\usage{
ts_gtrends_windows(
  keyword = NA,
  category = "0",
  geo = "CH",
  from = "2019-01-01",
  prevent_window_shrinkage = FALSE,
  stepsize = 7,
  windowsize = 80,
  n_windows = 12,
  quiet = FALSE,
  wait = 60,
  retry = 5
)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query
keywords. Contrary to \code{\link[gtrendsR:gtrends]{gtrendsR::gtrends()}}, if different keywords are used,
mulitple queries are sent to Google, and each series is individually
normalized.}

\item{category}{A character vector, listing the categories, defaults to “0”.}

\item{geo}{A character vector denoting geographic regions for the query, default is “CH”.}

\item{from}{Start of timeframe in YYYY-mm-dd form.}

\item{prevent_window_shrinkage}{If \code{TRUE}, ensures that the last time-window
is large enough to yield the desired frequency. Default is \code{FALSE}.}

\item{stepsize}{Number of days (integer) between the start days of the
respective time-windows.}

\item{windowsize}{Number of days (integer) from start date to end date of
each window.}

\item{n_windows}{Number (integer) of time-windows.}

\item{quiet}{If TRUE won't display messages related to server interactions. Default is FALSE.}

\item{wait}{Seconds to wait between attempts, where waiting time is \code{attempt * wait}.}

\item{retry}{Number of attempts, in case the query request does not succeed.}
}
\value{
A tibble of time series with columns \emph{window}, \emph{time}, \emph{value},
where \emph{window} is the time window
indicated by start and end date of the window.
}
\description{
Downloads approximate daily series over a sequence of time-windows by first
constructing the appropriate time windows and then downloading daily data
for each time-window using \code{\link[trendecon:ts_gtrends]{trendecon::ts_gtrends()}}.
}
\section{Notes}{

Time-windows may overlap - and will do so if \code{stepsize} < \code{windowsize}.
}

\seealso{
\code{\link[trendecon:ts_gtrends]{trendecon::ts_gtrends()}}
}
